pyproject.toml
[tool.poetry]
name = "madr"
version = "0.1.0"
description = "Meu Acervo Digital de Romances"
authors = ["Élysson MR <elyssonmr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.4.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
psycopg = {extras = ["binary"], version = "^3.2.1"}
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.8.2"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.5.6"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
factory-boy = "^3.3.0"
testcontainers = "^4.7.2"
pytest-asyncio = "^0.23.8"
taskipy = "^1.13.0"
freezegun = "^1.5.1"
gevent = "^24.2.1"
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_mode = 'auto'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev madr/app.py'
pre_test = 'task lint'
test = 'pytest -s --cov=madr --cov-report term-missing -vv'
post_test = 'coverage html'
init_migration = 'alembic init migrations'
migrate = 'alembic upgrade head'
migration_history = 'alembic history'
docker_build = 'docker build -t "madr" .'
